requirements.txt
streamlit
gTTS
PyPDF2
fpdf2